backend/funnyboard/src/main/java/funnyboard/service/ArticleService.java
package funnyboard.service;

import funnyboard.entity.Article;
import funnyboard.repository.ArticleRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ArticleService {

    private final ArticleRepository articleRepository;

    public ArticleService(ArticleRepository articleRepository) {
        this.articleRepository = articleRepository;
    }

    public List<Article> findAllArticles() {
        return articleRepository.findAll();
    }
}
